Move most of the packages to under internal
> Naming your package as internal, hides your package’s internals even
> more, brings more encapsulation.
