+++ cmd/docker-app/inspect.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/docker/app/renderer"
+	"github.com/docker/app/internal/renderer"
 	"github.com/docker/cli/cli"
 	"github.com/spf13/cobra"
 )
