+++ cmd/docker-app/render.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/docker/app/internal"
-	"github.com/docker/app/renderer"
+	"github.com/docker/app/internal/renderer"
 	"github.com/docker/cli/cli"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
