+++ cmd/docker-app/save.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/docker/app/packager"
+	"github.com/docker/app/internal/packager"
 	"github.com/docker/cli/cli"
 	"github.com/spf13/cobra"
 )
