+++ cmd/docker-app/split.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/docker/app/internal"
-	"github.com/docker/app/packager"
+	"github.com/docker/app/internal/packager"
 	"github.com/docker/cli/cli"
 	"github.com/spf13/cobra"
 )
