+++ internal/packager/split.go
