+++ internal/renderer/helm.go
@@ -8,12 +8,12 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/docker/app/packager"
-	"github.com/docker/app/templateconversion"
-	"github.com/docker/app/templateloader"
-	"github.com/docker/app/templatev1beta2"
-	"github.com/docker/app/types"
-	"github.com/docker/app/utils"
+	"github.com/docker/app/internal/packager"
+	"github.com/docker/app/internal/templateconversion"
+	"github.com/docker/app/internal/templateloader"
+	"github.com/docker/app/internal/templatev1beta2"
+	"github.com/docker/app/internal/types"
+	"github.com/docker/app/internal/utils"
 	conversion "github.com/docker/cli/cli/command/stack/kubernetes"
 	"github.com/docker/cli/cli/compose/loader"
 	"github.com/docker/cli/kubernetes/compose/v1beta2"
