+++ internal/templateloader/merge.go
